\name{demo_TEST}
\alias{demo_TEST}
\title{Demo, Batch Testing a GRNN}
\description{
simulates the grnn defined by C, W, & SIGMA and performs a batch test
}
\details{
performs batch-mode testing on the synthesized grnn defined by C, W, and SIGMA using test datafiles
}
\author{Arnold Arrington}
\keyword{misc}
