\name{mean_center_X}
\alias{mean_center_X}
\title{Perform Mean-centering on a Matrix}
\description{
return the mean-centered matrix and vector of variable (row) means
}
\usage{
mean_center_X(X)
}
\arguments{
  \item{X}{input matrix}
}
\value{
  \item{X_mc}{mean-centered matrix}
  \item{mn_Vect}{vector of variable means}
}
\author{Arnold Arrington}
\keyword{neural}
