\name{P_trn}
\docType{data}
\alias{P_trn}
\title{Training Pattern Data}
\description{
provides pattern data for the TRAIN demo
}
\usage{data(P_trn)}
\format{matrix, 500 observations (rows), 14 variables}
\source{PROBEN1, ftp.cs.cmu.edu /afs/cs/project/connect/bench/contrib/prechelt}
\referenced{
    Proben1 -- A Set of Neural Network
    Problems and Benchmarking Rules,
    Lutz Prechelt
    Fakultat fur Informatik
    Universitat Karlsruke
    1994
}
\keyword{datasets}
