\name{P_tst}
\docType{data}
\alias{P_tst}
\title{Test Pattern Data}
\description{
provides pattern data for the TEST demo
}
\usage{data(P_tst)}
\format{matrix, 1052 observations (rows), 14 variables}
\source{PROBEN1, ftp.cs.cmu.edu /afs/cs/project/connect/bench/contrib/prechelt.}
\referenced{
    Proben1 -- A Set of Neural Network
    Problems and Benchmarking Rules,
    Lutz Prechelt
    Fakultat fur Informatik
    Universitat Karlsruke
    1994
}
\keyword{datasets}
