\name{plot_Box}
\alias{plot_Box}
\title{Generate Error Box/Whisker Plot}
\description{
produce box/whisker plot for the absolute error measure between targets and grnn estimates
}
\usage{
plot_Box(T, T_hat, NOVAL=NA)
}
\arguments{
  \item{T}{targets}
  \item{T_hat}{neural net estimates}
  \item{NOVAL}{missing value parameter}
}
\details{
}
\author{Arnold Arrington}
\keyword{neural}
