\name{T_trn}
\docType{data}
\alias{T_trn}
\title{Training Target Data}
\description{
provides target data for the TRAIN demo
}
\usage{data(T_trn)}
\format{matrix, 500 target samples (rows), 3 output variables}
\source{PROBEN1, ftp.cs.cmu.edu /afs/cs/project/connect/bench/contrib/prechelt.}
\referenced{
    Proben1 -- A Set of Neural Network
    Problems and Benchmarking Rules,
    Lutz Prechelt
    Fakultat fur Informatik
    Universitat Karlsruke
    1994
}
\keyword{datasets}
