\name{var_scale_X}
\alias{var_scale_X}
\title{Perform Variance Scaling on a Matrix}
\description{
return the variance-scaled matrix and the vector of variable variances
}
\usage{
var_scale_X(X)
}
\arguments{
  \item{X}{input matrix}
}
\value{
  \item{X_var}{variance-scaled matrix}
  \item{var_Vect}{vector of variable variances}
}
\author{Arnold Arrington}
\keyword{neural}
